# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BrewDesk Café Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tables:
    get:
      operationId: listTables
      tags: [tables]
      summary: List all tables with current status
      responses:
        "200":
          description: List of tables
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Table"
    post:
      operationId: createTable
      tags: [tables]
      summary: Create a new table
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TableInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
  /tables/{id}:
    get:
      operationId: getTable
      tags: [tables]
      summary: Get table by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
    patch:
      operationId: updateTable
      tags: [tables]
      summary: Update a table
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TableUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
    delete:
      operationId: deleteTable
      tags: [tables]
      summary: Delete a table
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Rooms ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Table:
      type: object
      required: [id, number, capacity, status, type, posX, posY, hourlyRate]
      properties:
        id:
          type: integer
        number:
          type: integer
        capacity:
          type: integer
        status:
          type: string
          enum: [available, reserved, occupied]
        type:
          type: string
          enum: [regular, football, work]
        hourlyRate:
          type: number
          minimum: 0
        posX:
          type: integer
        posY:
          type: integer
        qrCode:
          type: ["string", "null"]
        minOrderAmount:
          type: ["number", "null"]
    TableInput:
      type: object
      required: [number, capacity, type]
      properties:
        number:
          type: integer
        capacity:
          type: integer
        type:
          type: string
          enum: [regular, football, work]
        hourlyRate:
          type: number
          minimum: 0
        posX:
          type: integer
        posY:
          type: integer
        minOrderAmount:
          type: number
    TableUpdate:
      type: object
      properties:
        number:
          type: integer
        capacity:
          type: integer
        status:
          type: string
          enum: [available, reserved, occupied]
        type:
          type: string
          enum: [regular, football, work]
        hourlyRate:
          type: number
          minimum: 0
        posX:
          type: integer
        posY:
          type: integer
        minOrderAmount:
          type: number

    # ── Rooms ───────────────────────────────────────────────────────────
...